Implement getting weather from last known location
Currently have a short Toast to confirm address but geocoder usage is UI blocking right now. Will implement better method for display in main activity.

+++ app/src/main/java/com/kodatos/cumulonimbus/apihelper/WeatherAPIService.java
@@ -17,12 +17,12 @@ public interface WeatherAPIService {
     Call<CurrentWeatherModel> getCurrentWeatherByString(@Query("q") String q, @Query("appid") String appid);
 
     @GET("data/2.5/weather")
-    Call<CurrentWeatherModel> getCurrentWeatherByCoords(@Query("lat") float lat, @Query("lon") float lon, @Query("appid") String appid);
+    Call<CurrentWeatherModel> getCurrentWeatherByCoords(@Query("lat") double lat, @Query("lon") double lon, @Query("appid") String appid);
 
     @GET("data/2.5/forecast")
     Call<ForecastWeatherModel> getForecastWeatherByString(@Query("q") String q, @Query("appid") String appid);
 
     @GET("data/2.5/forecast")
-    Call<ForecastWeatherModel> getForecastWeatherByCoords(@Query("lat") float lat, @Query("lon") float lon, @Query("appid") String appid);
+    Call<ForecastWeatherModel> getForecastWeatherByCoords(@Query("lat") double lat, @Query("lon") double lon, @Query("appid") String appid);
 
 }
